Create 2.5-Sweetwater.md
Treat this as incredibly tenuous and likely to disappear. Our recent Coinlist listing determined a time period that we need to launch a network in, and this is staking out the proposal of what that network will be. It could be that this becomes Oregon Trail, or Oregon Trail has functionality pushed out. Currently it’s undecided.