Remove references to maxWantedStake
As per a conversation with @vegaklaus, there is one single network
parameter, maxStakePerValidator, that dictates the highest amount of
stake that a single validator can have. This will default to a value
that ensures that at this maximum, validators will receive the optimum
reward. In later networks, it may be possible for validators to set
their own max.

- Remove max_wanted_stake references