Rename files to include codes
First stab. All filenames are being renamed so that they have a unique
ID based on their sequence (4 digit ID, unchanged) and a short,
hopefully representative enough code.

Initially I started making it:
`([0-9]{4})-([A-Z]{6})+` so that the text portion had to be 6
characters, but then I changed it to be a looser `([0-9]{4})-([A-Z])+`
to allow a more expressive text portion, given that the point of
adding this text code is to make it referenceable and make sense for
humans. Do we accomplish that? If you saw a test refer to
`0038-ORDERLIQ-02` would you know it meant Acceptance Criteria 2 on
the Liquidity Order spec?

- Rename all files