chore: update links and AC codes
- Links to proto files updated

- 0040-ASSF-002 - Marked as NOT for trading enabled.

NOTE: in cosmic could we 'lock' an asset in the first instance. Then decide about full removal in the future? If this is not the case (or in the very distant future we can remove this for now)

No other references to removing an asset have been found in the text of the spec.