Update 6001-SORD-submit_orders.md
A lot of this feels quite specific to the solution, but it also might not be worth changing this now if this is expected behaviour based on how users expect an exchange to work? 

For example:
input field pre-populated -> ... have the previous value for the selected market available e.g. pre-populated
hit up/down on the keyboard to increase -> ... be able to quickly tweak the size by the market's min-contract size e.g. hit up/down on the keyboard to increase