User interface AC tweaks (#1216)
* Update 0001-WALL-wallet.md

Small tweak to make the wording of downloading the wallet from .xyz less solution specific (we might need to change the solution there)

* Update 0002-WCON-connect_vega_wallet.md

Added - **should** have a way to easily access help should I need to troubleshoot my connection e.g. link to the docs (needs a link attached to it)

* Update 6001-SORD-submit_orders.md

A lot of this feels quite specific to the solution, but it also might not be worth changing this now if this is expected behaviour based on how users expect an exchange to work? 

For example:
input field pre-populated -> ... have the previous value for the selected market available e.g. pre-populated
hit up/down on the keyboard to increase -> ... be able to quickly tweak the size by the market's min-contract size e.g. hit up/down on the keyboard to increase