feat: removal of checkpoint
It has been decided that the checkpoint (LNL) feature will be removed from the core as this is no longer a viable option for restarting the network.

In its place will be snapshots and any required tooling.

When reviewing this PR please take a look at the following issues (specifically the QA issue):

- https://github.com/vegaprotocol/system-tests/issues/3346
- https://github.com/vegaprotocol/vega/issues/10506