feat: Add extra validator ranking reward metric AC
Adds an extra validator ranking reward metric AC:

For reward metrics relating to trading, an individual must meet the staking requirement AND notional time-weighted average position requirement) set in the recurring transfer. If they do not then their reward metric is set to 0. Note, these requirements do not apply to the validator ranking metric or the market creation reward metric.

e.g. For a party that is a consensus or standby validator, the staking requirement and notional time-weighted average position requirement do not apply to their validator ranking metric